Changed Player's bullets to Slashes.
Textures for Player's bullet are changed to Slash textures. When the player left clicks, they Slash their sword and also send out a Slash Bullet. When the Boss gets hit they flash for 1 frame white. Added damage text that spawns when the Player is hit, when the Boss's bullets are hit and when the Boss gets hit.